As a user, I want to see my previous executions and their result. [Done] feat: Display previous executions with their result of a registered user.

+++ app/Http/Controllers/GoController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
 use Illuminate\View\View;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
@@ -10,7 +11,8 @@ use Illuminate\Support\Facades\Storage;
 
 class GoController extends Controller
 {
-    public function create() : View {
+    public function create(): View
+    {
         $formAction = route('execute-go');
         return view('execution.goCode', compact('formAction'));
     }
@@ -20,15 +22,23 @@ class GoController extends Controller
         $file_path = 'tmp/' . uniqid() . '.go';
         $full_path = Storage::path($file_path);
         ($request->input) ? Storage::put($file_path, $request->input) : Storage::put($file_path, '');
-        $result = new Process([env('GO_PATH'),'run', $full_path], null, ['GOCACHE' => env('GOCACHE_PATH')]);
+        $result = new Process([env('GO_PATH'), 'run', $full_path], null, ['GOCACHE' => env('GOCACHE_PATH')]);
         $result->run();
         $output = $result->isSuccessful() ? $result->getOutput() : $result->getErrorOutput();
-        Storage::delete($file_path);
-        return view('execution.goCode',[
+        if (auth()->check()) {
+            User::find(auth()->id())->submissions()->create([
+                'user_id' => auth()->id(),
+                'input' => $file_path,
+                'output' => $result->getOutput(),
+                'language' => "go",
+            ]);
+        } else {
+            Storage::delete($file_path);
+        }
+        return view('execution.goCode', [
             'formAction' => route('execute-go'),
             'input' => $request->input,
             'output' => $output,
         ]);
     }
-
 }
